Analysing the network

Calculates some indicators for each node in the network.

analyze_network(Omega,nv,...)

Arguments

Omega
a network object

nv
the level of cutoff at which the analysis should be done

…
label_v : (optionnal) name of the genes

Value

A matrix containing, for each node, its betweenness,its degree, its output, its closeness.

References

Jung, N., Bertrand, F., Bahram, S., Vallat, L., and Maumy-Bertrand, M. (2013). Cascade: a R-package to study, predict and simulate the diffusion of a signal through a temporal gene network. Bioinformatics, btt705.

Vallat, L., Kemper, C. A., Jung, N., Maumy-Bertrand, M., Bertrand, F., Meyer, N., ... & Bahram, S. (2013). Reverse-engineering the genetic circuitry of a cancer cell with predicted intervention in chronic lymphocytic leukemia. Proceedings of the National Academy of Sciences, 110(2), 459-464.

Examples

        data(network)
        analyze_network(network,nv=0)

#> Loading required package: tnet

#> Loading required package: igraph

#> 
#> Attaching package: ‘igraph’

#> The following object is masked from ‘package:Cascade’:
#> 
#>     compare

#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum

#> The following object is masked from ‘package:base’:
#> 
#>     union

#> Loading required package: survival

#> tnet: Analysis of Weighted, Two-mode, and Longitudinal networks.
#> Type ?tnet for help.

#>     node betweenness degree      output    closeness
#> 1      1           0     41 2.115925334  53.79741935
#> 2      2          41     39 2.513381538  49.64857375
#> 3      3           0     32 3.103278157  84.46040797
#> 4      4           0     52 4.837819109 133.68995032
#> 5      5           0     35 1.907441576  53.13521459
#> 6      6           0     35 1.204158962  49.57296793
#> 7      7           0     29 0.491259265  22.02539627
#> 8      8           0     22 0.326333564  67.50590184
#> 9      9         256     37 2.141792871  44.25782447
#> 10    10         137     29 2.173037334  46.65947816
#> 11    11           0     42 2.117408416 105.90041992
#> 12    12           0     46 3.933610474  92.85801386
#> 13    13           0     38 0.540728340  18.54085205
#> 14    14           0     52 1.562246739  52.31739199
#> 15    15           0     42 3.202031415 109.39718660
#> 16    16         281     43 7.707581233 155.29444082
#> 17    17           0     38 3.122945751  78.88975647
#> 18    18           0     44 3.231599597  83.92527638
#> 19    19           0     32 0.522620143  14.85714028
#> 20    20           0     36 0.336035752  20.08353066
#> 21    21           0      3 0.086393474   7.48111170
#> 22    22           0      5 0.065569380  10.29054106
#> 23    23           0     13 0.268899273   9.16543299
#> 24    24           0      2 0.041242098   6.35774369
#> 25    25           0      2 0.037751446   1.34501648
#> 26    26           0     18 0.555031988  20.11288248
#> 27    27           0     42 3.472742236  82.32413143
#> 28    28           0      9 0.550144384  25.36067274
#> 29    29           0     13 0.236333735   7.47709997
#> 30    30           0      7 0.198416275  18.22501773
#> 31    31           0     21 0.549116529  13.74463021
#> 32    32           0      9 0.157283725  10.97812862
#> 33    33           0     11 0.208701671  17.48525682
#> 34    34           0     11 0.521664153  17.87358603
#> 35    35           0     28 2.266892032  52.48268010
#> 36    36           0      1 0.004276882   1.47467352
#> 37    37           0     11 0.430166932  17.93626676
#> 38    38           0      1 0.004096049   0.07445144
#> 39    39           0      6 0.064282146   1.69417015
#> 40    40           0     16 0.462519609  28.27350339
#> 41    41           0      2 0.007001813   0.15809376
#> 42    42          87     20 0.622340946  16.17173024
#> 43    43           0      0 0.000000000   0.00000000
#> 44    44          39     14 0.186791633   5.37542938
#> 45    45          13      4 0.055697064   2.51165809
#> 46    46           0      2 0.007447328   0.13536563
#> 47    47          46     13 0.240128861   6.51734262
#> 48    48         115     17 0.693828497  15.33683233
#> 49    49          51     15 0.378445509   9.29640433
#> 50    50           2      2 0.139848600   2.54194448
#> 51    51           6      2 0.024463423   1.58268047
#> 52    52           0      0 0.000000000   0.00000000
#> 53    53          12      8 0.115966349   2.49617498
#> 54    54           0      0 0.000000000   0.00000000
#> 55    55           0      0 0.000000000   0.00000000
#> 56    56           0      1 0.016258068   0.29551319
#> 57    57          10      4 0.494046644   8.97999078
#> 58    58           0      0 0.000000000   0.00000000
#> 59    59           0      1 0.002284263   0.04151969
#> 60    60           6      3 0.092855604   1.68778085
#> 61    61           6      1 0.113674696   2.06619706
#> 62    62           9      1 0.077398170   1.40682031
#> 63    63          24      3 0.187822345   3.41393459
#> 64    64           0      0 0.000000000   0.00000000
#> 65    65           2      1 0.026431811   0.48043525
#> 66    66           0      0 0.000000000   0.00000000
#> 67    67           0      0 0.000000000   0.00000000
#> 68    68          34      5 0.124270366   2.25878823
#> 69    69           0      0 0.000000000   0.00000000
#> 70    70           0      0 0.000000000   0.00000000
#> 71    71           6      4 0.071416296   1.29809136
#> 72    72          12      2 0.241063855   4.38167372
#> 73    73           0      0 0.000000000   0.00000000
#> 74    74           3      4 0.152181171   2.76610625
#> 75    75          22      2 0.215218316   3.91189479
#> 76    76           7      2 0.066261991   1.20440464
#> 77    77           2      5 0.187968919   3.41659877
#> 78    78           4      2 0.135639956   2.46544647
#> 79    79          30      7 0.556373920  10.11287644
#> 80    80           0      0 0.000000000   0.00000000
#> 81    81           0      0 0.000000000   0.00000000
#> 82    82          30      2 0.612689152  11.13648478
#> 83    83          32      3 0.342675164   6.22860179
#> 84    84          12      4 0.279803534   5.08582172
#> 85    85          11      3 0.182800332   3.32265245
#> 86    86           3      3 0.100248936   1.82216503
#> 87    87           7      2 0.136366661   2.47865534
#> 88    88           0      0 0.000000000   0.00000000
#> 89    89           0      1 0.011706142   0.21277554
#> 90    90          27      4 0.256810158   4.66788487
#> 91    91           5      2 0.052450371   0.95335907
#> 92    92           5      2 0.032030132   0.58219257
#> 93    93          10      2 0.207746410   3.77608241
#> 94    94           5      3 0.103023466   1.87259601
#> 95    95           0      0 0.000000000   0.00000000
#> 96    96           0      0 0.000000000   0.00000000
#> 97    97           1      2 0.011715032   0.21293714
#> 98    98           1      1 0.003691672   0.06710132
#> 99    99           7      3 0.083935163   1.52563932
#> 100  100           0      0 0.000000000   0.00000000
#> 101  101           1      1 0.001458042   0.02650196
#> 102  102          13      6 0.214365366   3.89639123