Analysing the network

Calculates some indicators for each node in the network.

Usage

# S4 method for class 'network'
analyze_network(Omega, nv, label_v = NULL)

Arguments

Omega

a network object

nv

the level of cutoff at which the analysis should be done

label_v

(optionnal) the name of the genes

Value

A matrix containing, for each node, its betweenness,its degree, its output, its closeness.

References

Jung, N., Bertrand, F., Bahram, S., Vallat, L., and Maumy-Bertrand, M. (2014). Cascade: a R-package to study, predict and simulate the diffusion of a signal through a temporal gene network. Bioinformatics, btt705.

Vallat, L., Kemper, C. A., Jung, N., Maumy-Bertrand, M., Bertrand, F., Meyer, N., ... & Bahram, S. (2013). Reverse-engineering the genetic circuitry of a cancer cell with predicted intervention in chronic lymphocytic leukemia. Proceedings of the National Academy of Sciences, 110(2), 459-464.

Author

Nicolas Jung, Frédéric Bertrand , Myriam Maumy-Bertrand.

Examples

  data(network)
  analyze_network(network,nv=0)
#> Loading required package: tnet
#> Loading required package: igraph
#> 
#> Attaching package: ‘igraph’
#> The following object is masked from ‘package:Cascade’:
#> 
#>     compare
#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum
#> The following object is masked from ‘package:base’:
#> 
#>     union
#> Loading required package: survival
#> tnet: Analysis of Weighted, Two-mode, and Longitudinal networks.
#> Type ?tnet for help.
#>    node betweenness degree      output    closeness
#> 1     1           0     52 2.493277168 110.83148964
#> 2     2           0     51 1.873353046  91.33091185
#> 3     3           0     48 0.869999868  75.92040114
#> 4     4           0     44 1.712495042  82.94726787
#> 5     5           0     44 0.811053103  34.64847877
#> 6     6           0     51 5.326562329 152.65540461
#> 7     7           0     50 3.448573704 104.41478218
#> 8     8           0     49 1.086343159  32.84110423
#> 9     9           0     52 0.781594709  35.17592984
#> 10   10          23     47 2.193224490  59.86437756
#> 11   11         156     48 3.844765796  92.87430047
#> 12   12          57     35 1.356068126  38.80420637
#> 13   13         171     44 5.060737204 125.32045602
#> 14   14           7     43 0.991895840  31.47338043
#> 15   15           0      5 0.125014915   2.54460894
#> 16   16           1      9 0.073129838   1.48851710
#> 17   17          51     13 0.862046510  17.54647642
#> 18   18          18      9 0.365414532   7.43780922
#> 19   19           0      4 0.058039485   1.18136139
#> 20   20           4      9 0.299858554   6.10345380
#> 21   21           4      8 0.220877759   4.49584372
#> 22   22          20     13 0.370169895   7.53460197
#> 23   23           3      4 0.121253333   2.46804404
#> 24   24           5      9 0.238911191   4.86290418
#> 25   25           0      0 0.000000000   0.00000000
#> 26   26           4     12 0.200293631   4.07686527
#> 27   27          15     15 0.676393027  13.76760322
#> 28   28          20      7 0.386360480   7.86415231
#> 29   29           5      2 0.099861827   2.03263185
#> 30   30         104     18 1.831171079  37.27246708
#> 31   31           2      5 0.089213272   1.81588645
#> 32   32          13     13 0.311496789   6.34034358
#> 33   33           0     18 0.152389258   3.10179844
#> 34   34          15     13 0.580194498  11.80953575
#> 35   35           0      0 0.000000000   0.00000000
#> 36   36           0      0 0.000000000   0.00000000
#> 37   37           0      0 0.000000000   0.00000000
#> 38   38           0      0 0.000000000   0.00000000
#> 39   39           0      0 0.000000000   0.00000000
#> 40   40           0      0 0.000000000   0.00000000
#> 41   41           0      0 0.000000000   0.00000000
#> 42   42           0      0 0.000000000   0.00000000
#> 43   43           0      0 0.000000000   0.00000000
#> 44   44           0      0 0.000000000   0.00000000
#> 45   45           0      0 0.000000000   0.00000000
#> 46   46           0      0 0.000000000   0.00000000
#> 47   47           0      0 0.000000000   0.00000000
#> 48   48           0      0 0.000000000   0.00000000
#> 49   49           0      0 0.000000000   0.00000000
#> 50   50           0      0 0.000000000   0.00000000
#> 51   51           0      0 0.000000000   0.00000000
#> 52   52           0      0 0.000000000   0.00000000
#> 53   53           0      0 0.000000000   0.00000000
#> 54   54           0      0 0.000000000   0.00000000
#> 55   55           0     40 3.884171706 143.68410789
#> 56   56          28     30 1.213962675  34.79491085
#> 57   57          17      8 0.381344058   7.76204583
#> 58   58           6     15 0.839509022  23.76967274
#> 59   59           2      7 0.191033515   3.88838076
#> 60   60           0     22 0.590567211  73.18774872
#> 61   61           0     25 1.624415450  82.84061811
#> 62   62           0      9 0.086465686   1.75996086
#> 63   63           0      1 0.001638944   0.03335979
#> 64   64           7     26 1.344245113  37.33582438
#> 65   65           0     13 1.368119564  95.96946550
#> 66   66          25      9 0.505402921  10.28719487
#> 67   67          24      9 0.756855631  15.40537469
#> 68   68           0     11 0.175883203  12.74214052
#> 69   69           1      2 0.074150206   1.50928614
#> 70   70          10     34 1.050511028  32.75663861
#> 71   71          70     35 3.333568905  86.26612955
#> 72   72           0      8 0.211764898   4.31035651
#> 73   73           7      6 0.288167148   5.86548174
#> 74   74          11     10 0.352721778   7.17945528