Returns the position of edges in the network

Returns the position of edges in the network

Usage

# S4 method for class 'network'
position(net, nv = 0)

Arguments

net

a network object

nv

the level of cutoff at which the analysis should be done

Methods

list("signature(net = \"network\")")

Returns a matrix with the position of the node. This matrix can then be used as an argument in the plot function.

Examples

data(Net)
position(Net)
#>       nom                  
#>  [1,]   1 77.46997 156.9371
#>  [2,]   2 73.41193 148.3426
#>  [3,]   3 78.20901 153.4573
#>  [4,]   4 74.85340 147.6659
#>  [5,]   5 83.44547 145.6638
#>  [6,]   6 80.35849 155.5171
#>  [7,]   7 75.86694 154.7806
#>  [8,]   8 81.57897 156.3231
#>  [9,]   9 75.98990 156.6662
#> [10,]  10 78.31870 149.8716
#> [11,]  11 78.34996 150.8620
#> [12,]  12 80.95375 148.3444
#> [13,]  14 82.59770 157.4199
#> [14,]  16 79.75868 148.3432
#> [15,]  17 83.61067 153.8689
#> [16,]  18 82.37833 156.2044
#> [17,]  20 77.44993 157.9903
#> [18,]  21 75.35504 157.3956
#> [19,]  22 75.88524 144.0366
#> [20,]  23 76.84126 154.9939
#> [21,]  24 77.14492 151.9194
#> [22,]  25 72.08948 146.1022
#> [23,]  26 75.59825 150.5932
#> [24,]  27 78.31967 157.0885
#> [25,]  28 77.06044 155.6108
#> [26,]  29 79.14918 157.7473
#> [27,]  30 76.53113 154.5199
#> [28,]  31 75.41251 149.2612
#> [29,]  32 80.50145 151.0859
#> [30,]  33 78.29143 147.6480
#> [31,]  34 73.28734 147.2623
#> [32,]  35 77.25397 148.2302
#> [33,]  36 81.75616 152.6557
#> [34,]  37 82.06085 146.2995
#> [35,]  38 75.21594 154.3973
#> [36,]  39 82.81504 154.1239
#> [37,]  40 83.43239 144.1633
#> [38,]  41 83.56811 155.8781
#> [39,]  42 81.35389 154.0321
#> [40,]  43 82.28192 158.2915
#> [41,]  44 75.08648 145.4784
#> [42,]  45 79.32739 149.1486
#> [43,]  46 84.12278 150.4945
#> [44,]  47 80.91046 152.7952
#> [45,]  48 84.29274 156.2554
#> [46,]  49 79.73959 158.0127
#> [47,]  50 80.24042 156.2906
#> [48,]  51 79.25495 146.3922
#> [49,]  52 83.56600 153.0075
#> [50,]  53 81.83580 151.8384
#> [51,]  54 84.32090 151.7896
#> [52,]  55 77.27711 150.8059
#> [53,]  56 83.47931 151.9744
#> [54,]  57 79.09491 148.0179
#> [55,]  58 78.57701 155.2934
#> [56,]  59 81.25469 151.5470
#> [57,]  60 79.72995 148.8995
#> [58,]  61 85.07331 152.1485
#> [59,]  62 74.79461 155.9138
#> [60,]  63 73.96161 152.7643
#> [61,]  64 80.50686 154.2330
#> [62,]  65 79.82471 150.4433
#> [63,]  66 82.80514 151.1132
#> [64,]  67 80.38765 150.5509
#> [65,]  68 74.94081 153.0034
#> [66,]  69 79.80670 151.9585
#> [67,]  70 77.63792 155.4033
#> [68,]  71 82.08996 147.9185
#> [69,]  72 77.60293 152.9320
#> [70,]  73 79.68958 152.5670
#> [71,]  74 82.76072 150.0575
#> [72,]  75 86.28499 150.0978
#> [73,]  76 86.36585 150.9036
#> [74,]  77 85.57841 148.7860
#> [75,]  78 75.76539 151.3881
#> [76,]  79 75.91043 152.7741
#> [77,]  80 78.93664 150.0503
#> [78,]  81 86.14024 149.3583
#> [79,]  82 83.50836 148.3274
#> [80,]  83 73.38851 150.0830
#> [81,]  84 72.05803 152.6488
#> [82,]  85 81.84141 149.3441
#> [83,]  86 82.59948 149.4024
#> [84,]  87 79.50717 154.7953
#> [85,]  88 81.71096 150.7609
#> [86,]  89 77.06174 149.9452
#> [87,]  90 76.57152 148.1705
#> [88,]  91 77.23058 149.2268
#> [89,]  92 74.70478 150.4030
#> [90,]  93 80.33163 145.9464
#> [91,]  94 85.75209 153.9310
#> [92,]  95 82.77499 152.8827
#> [93,]  96 84.61277 144.8705
#> [94,]  97 72.25790 153.6752
#> [95,]  98 72.76612 145.3700
#> [96,]  99 76.83267 143.2486
#> [97,] 100 81.29842 145.8933