Returns the position of edges in the network

Returns the position of edges in the network

Methods

signature(net = "network")

Returns a matrix with the position of the node. This matrix can then be used as an argument in the plot function.

Examples

data(Net)
position(Net)

#>       nom                  
#>  [1,]   1 86.28886 140.8559
#>  [2,]   2 74.76234 143.9413
#>  [3,]   3 82.00023 140.6807
#>  [4,]   4 76.59251 145.4013
#>  [5,]   5 82.69152 150.1110
#>  [6,]   6 84.64301 141.4651
#>  [7,]   7 83.28489 138.5121
#>  [8,]   8 85.69518 146.1499
#>  [9,]   9 84.11061 136.8050
#> [10,]  10 79.76102 144.6050
#> [11,]  11 79.64630 145.6379
#> [12,]  12 84.08473 145.4271
#> [13,]  14 83.74666 149.0822
#> [14,]  16 79.97621 141.1700
#> [15,]  17 84.03095 146.0752
#> [16,]  18 81.60416 149.4737
#> [17,]  20 86.99650 138.8308
#> [18,]  21 86.40023 137.8015
#> [19,]  22 76.42021 149.8191
#> [20,]  23 79.06276 145.8588
#> [21,]  24 81.63390 140.5510
#> [22,]  25 88.50194 150.5046
#> [23,]  26 79.87445 138.9959
#> [24,]  27 85.05660 138.5914
#> [25,]  28 84.95835 139.4037
#> [26,]  29 86.74924 139.9620
#> [27,]  30 84.27312 139.9856
#> [28,]  31 76.64580 144.7096
#> [29,]  32 80.50715 143.1912
#> [30,]  33 77.41636 143.7088
#> [31,]  34 74.67060 144.8507
#> [32,]  35 78.81372 144.3508
#> [33,]  36 83.76781 143.7021
#> [34,]  37 85.23235 147.6617
#> [35,]  38 82.73252 137.7640
#> [36,]  39 81.91081 148.1621
#> [37,]  40 83.45251 151.7663
#> [38,]  41 85.73732 147.2449
#> [39,]  42 84.23519 142.8000
#> [40,]  43 85.69105 148.7360
#> [41,]  44 76.28171 148.0160
#> [42,]  45 79.56361 147.4023
#> [43,]  46 86.48614 144.6473
#> [44,]  47 82.97719 145.1906
#> [45,]  48 84.53249 148.4895
#> [46,]  49 87.02899 143.4204
#> [47,]  50 81.55349 147.3421
#> [48,]  51 80.85750 149.0865
#> [49,]  52 81.34495 146.3667
#> [50,]  53 83.02778 143.5723
#> [51,]  54 86.08921 143.7048
#> [52,]  55 79.14095 141.0294
#> [53,]  56 85.33512 145.3948
#> [54,]  57 81.69051 143.9616
#> [55,]  58 82.44720 138.9503
#> [56,]  59 79.96968 149.2808
#> [57,]  60 78.16365 142.5321
#> [58,]  61 88.37430 143.1373
#> [59,]  62 84.49551 137.4855
#> [60,]  63 76.97263 145.9947
#> [61,]  64 82.46222 144.2710
#> [62,]  65 81.98906 142.9360
#> [63,]  66 78.36303 141.6996
#> [64,]  67 82.57339 143.3169
#> [65,]  68 81.46935 137.9421
#> [66,]  69 81.47243 145.1343
#> [67,]  70 79.92819 142.9488
#> [68,]  71 82.23993 146.6593
#> [69,]  72 81.29512 139.9071
#> [70,]  73 84.77261 141.9546
#> [71,]  74 77.74168 141.7818
#> [72,]  75 89.05995 145.0358
#> [73,]  76 88.48701 145.4172
#> [74,]  77 88.48639 146.2140
#> [75,]  78 79.11044 143.4273
#> [76,]  79 80.65854 138.5652
#> [77,]  80 79.55240 141.5778
#> [78,]  81 88.82702 144.2181
#> [79,]  82 86.59666 145.8327
#> [80,]  83 75.43997 146.0000
#> [81,]  84 75.69101 147.4459
#> [82,]  85 84.33055 144.3317
#> [83,]  86 84.89989 143.4359
#> [84,]  87 83.36160 140.3196
#> [85,]  88 83.25818 142.1160
#> [86,]  89 80.43235 140.3546
#> [87,]  90 78.51154 143.4604
#> [88,]  91 81.22866 141.8498
#> [89,]  92 77.45031 139.3441
#> [90,]  93 82.91702 148.0656
#> [91,]  94 79.99389 148.2784
#> [92,]  95 87.47846 141.8726
#> [93,]  96 82.34090 151.8975
#> [94,]  97 75.07261 146.8726
#> [95,]  98 87.76647 151.1557
#> [96,]  99 77.01369 151.0005
#> [97,] 100 86.43040 147.0344