Returns the position of edges in the network

Returns the position of edges in the network

Usage

# S4 method for class 'network'
position(net, nv = 0)

Arguments

net

a network object

nv

the level of cutoff at which the analysis should be done

Methods

list("signature(net = \"network\")")

Returns a matrix with the position of the node. This matrix can then be used as an argument in the plot function.

Examples

data(Net)
position(Net)
#>       nom                  
#>  [1,]   1 77.62579 157.0584
#>  [2,]   2 75.86162 145.5901
#>  [3,]   3 77.29868 153.1147
#>  [4,]   4 74.31050 147.6802
#>  [5,]   5 83.89926 146.1839
#>  [6,]   6 80.05434 155.1712
#>  [7,]   7 76.45764 155.5151
#>  [8,]   8 82.02396 155.7621
#>  [9,]   9 75.57764 156.6670
#> [10,]  10 78.35228 149.7702
#> [11,]  11 79.85832 148.8184
#> [12,]  12 80.89050 148.5299
#> [13,]  14 81.44963 157.3230
#> [14,]  16 77.57833 149.8369
#> [15,]  17 83.20596 153.8566
#> [16,]  18 81.16200 155.9456
#> [17,]  20 77.76132 158.0188
#> [18,]  21 76.00897 157.7624
#> [19,]  22 74.01613 144.5360
#> [20,]  23 75.62523 154.4312
#> [21,]  24 76.64469 152.2593
#> [22,]  25 79.10569 142.6129
#> [23,]  26 74.73178 151.1691
#> [24,]  27 77.97744 156.6956
#> [25,]  28 77.21315 155.6830
#> [26,]  29 79.24969 157.6255
#> [27,]  30 76.38985 154.9049
#> [28,]  31 77.84272 146.4079
#> [29,]  32 80.27481 151.3799
#> [30,]  33 77.92357 147.9329
#> [31,]  34 74.26854 146.2650
#> [32,]  35 76.61424 148.4267
#> [33,]  36 81.39357 151.9091
#> [34,]  37 81.55097 146.1946
#> [35,]  38 74.74029 154.6516
#> [36,]  39 82.31326 153.8545
#> [37,]  40 85.20149 145.5326
#> [38,]  41 83.91929 155.6658
#> [39,]  42 80.93662 153.9012
#> [40,]  43 82.58841 157.7389
#> [41,]  44 73.44880 145.9081
#> [42,]  45 80.09926 148.0367
#> [43,]  46 84.03046 150.8273
#> [44,]  47 81.15431 152.2193
#> [45,]  48 83.24490 156.3656
#> [46,]  49 80.13362 157.5086
#> [47,]  50 79.06117 155.9157
#> [48,]  51 79.89945 145.6803
#> [49,]  52 83.28830 153.1798
#> [50,]  53 81.82004 151.8780
#> [51,]  54 83.65226 150.5278
#> [52,]  55 76.96182 150.6762
#> [53,]  56 83.83091 152.0310
#> [54,]  57 78.82390 148.4193
#> [55,]  58 77.52936 154.7118
#> [56,]  59 81.65878 150.7512
#> [57,]  60 78.86766 149.5199
#> [58,]  61 84.79111 152.5720
#> [59,]  62 74.79204 156.1215
#> [60,]  63 73.36673 152.0233
#> [61,]  64 79.87701 154.1460
#> [62,]  65 79.16445 150.5090
#> [63,]  66 82.67888 150.8210
#> [64,]  67 79.98753 151.0154
#> [65,]  68 74.30660 153.1945
#> [66,]  69 81.17450 149.4634
#> [67,]  70 75.02532 153.7643
#> [68,]  71 82.19025 148.3039
#> [69,]  72 77.84696 153.4190
#> [70,]  73 79.47406 152.5963
#> [71,]  74 82.67573 151.8654
#> [72,]  75 86.32603 150.7217
#> [73,]  76 86.19493 149.9689
#> [74,]  77 86.24802 151.5261
#> [75,]  78 75.55672 151.0400
#> [76,]  79 75.94398 153.1528
#> [77,]  80 78.12836 151.0402
#> [78,]  81 85.75086 149.3029
#> [79,]  82 83.50015 148.6658
#> [80,]  83 72.89340 149.5794
#> [81,]  84 71.64668 151.5817
#> [82,]  85 81.85064 149.5437
#> [83,]  86 82.20630 150.0122
#> [84,]  87 78.76255 154.4276
#> [85,]  88 80.58430 150.3077
#> [86,]  89 76.19552 150.2142
#> [87,]  90 76.11220 148.2913
#> [88,]  91 76.75863 149.6426
#> [89,]  92 74.65087 149.6649
#> [90,]  93 80.68069 146.1115
#> [91,]  94 85.41577 154.2702
#> [92,]  95 82.44209 153.1057
#> [93,]  96 84.19332 144.7133
#> [94,]  97 71.70009 152.6738
#> [95,]  98 80.07012 142.6054
#> [96,]  99 74.93735 143.7101
#> [97,] 100 82.25926 146.5941