Returns the position of edges in the network

Returns the position of edges in the network

# S4 method for network
position(net, nv = 0)

Arguments

net
a network object

nv
the level of cutoff at which the analysis should be done

Methods

list("signature(net = \"network\")")

Returns a matrix with the position of the node. This matrix can then be used as an argument in the plot function.

Examples

data(Net)
position(Net)

#>       nom                  
#>  [1,]   1 78.85484 152.7683
#>  [2,]   2 80.73265 140.3326
#>  [3,]   3 76.57021 147.8938
#>  [4,]   4 77.30162 141.0465
#>  [5,]   5 71.00997 147.0415
#>  [6,]   6 78.20188 150.8659
#>  [7,]   7 78.42931 149.5879
#>  [8,]   8 77.16807 151.5203
#>  [9,]   9 81.75499 152.7064
#> [10,]  10 77.48446 144.2810
#> [11,]  11 79.35835 144.5039
#> [12,]  12 77.06034 147.8797
#> [13,]  14 73.26918 151.5965
#> [14,]  16 80.48012 143.9473
#> [15,]  17 77.19330 149.4140
#> [16,]  18 71.89331 149.6357
#> [17,]  20 78.25099 153.8567
#> [18,]  21 79.26495 153.8774
#> [19,]  22 76.84748 138.4932
#> [20,]  23 72.35936 146.3824
#> [21,]  24 79.93850 147.4140
#> [22,]  25 71.58754 153.9699
#> [23,]  26 82.32988 145.5563
#> [24,]  27 80.09718 152.8794
#> [25,]  28 78.56182 151.4335
#> [26,]  29 77.59028 153.1586
#> [27,]  30 80.00614 150.6119
#> [28,]  31 80.72523 142.0140
#> [29,]  32 76.52055 145.8798
#> [30,]  33 77.83139 142.1685
#> [31,]  34 79.15973 140.0066
#> [32,]  35 79.11818 142.8366
#> [33,]  36 79.95207 148.2620
#> [34,]  37 74.97872 149.7272
#> [35,]  38 81.22257 150.4959
#> [36,]  39 73.98454 148.5653
#> [37,]  40 69.48633 146.5341
#> [38,]  41 76.24484 151.7853
#> [39,]  42 77.76353 149.8644
#> [40,]  43 74.82341 152.7588
#> [41,]  44 76.73467 139.5315
#> [42,]  45 76.14032 144.3769
#> [43,]  46 81.63763 148.3505
#> [44,]  47 79.24698 147.9431
#> [45,]  48 74.83840 151.3479
#> [46,]  49 77.23896 153.7782
#> [47,]  50 72.89956 149.2259
#> [48,]  51 74.07417 144.5736
#> [49,]  52 74.80053 147.9126
#> [50,]  53 78.35582 147.7615
#> [51,]  54 82.39085 148.9637
#> [52,]  55 75.46003 144.8044
#> [53,]  56 81.90065 147.7694
#> [54,]  57 78.17523 145.4298
#> [55,]  58 74.75948 149.1277
#> [56,]  59 73.89493 145.9768
#> [57,]  60 77.66456 143.5201
#> [58,]  61 83.11713 150.3445
#> [59,]  62 80.60282 152.1739
#> [60,]  63 73.37547 143.6575
#> [61,]  64 75.65534 148.7750
#> [62,]  65 79.66686 146.0697
#> [63,]  66 75.47854 143.4149
#> [64,]  67 79.00442 146.3063
#> [65,]  68 81.55500 149.3232
#> [66,]  69 81.08489 145.7259
#> [67,]  70 73.59143 147.3852
#> [68,]  71 74.16051 146.8806
#> [69,]  72 78.02293 147.1075
#> [70,]  73 79.37085 149.7686
#> [71,]  74 74.77069 144.0244
#> [72,]  75 84.26284 148.0404
#> [73,]  76 83.97207 149.5187
#> [74,]  77 84.07898 148.7472
#> [75,]  78 76.70146 145.0884
#> [76,]  79 80.23954 145.9696
#> [77,]  80 78.74175 144.8752
#> [78,]  81 83.99918 147.3008
#> [79,]  82 79.62773 148.8688
#> [80,]  83 74.81738 141.4505
#> [81,]  84 71.97613 142.2116
#> [82,]  85 80.09548 146.8718
#> [83,]  86 82.28782 146.5345
#> [84,]  87 76.05275 149.6222
#> [85,]  88 81.39892 146.8329
#> [86,]  89 82.05737 144.4835
#> [87,]  90 76.80720 143.0073
#> [88,]  91 79.96052 145.1206
#> [89,]  92 73.90196 142.7808
#> [90,]  93 75.03531 146.1846
#> [91,]  94 72.28061 148.2045
#> [92,]  95 81.86576 151.0836
#> [93,]  96 69.48840 147.7425
#> [94,]  97 71.45098 143.1290
#> [95,]  98 70.89808 153.2679
#> [96,]  99 77.86785 138.2195
#> [97,] 100 75.68637 150.1935